moved unit tests to folder
